moved Configuration up the package structure so it can be shared
